+++ src/test/java/org/apiphany/client/ContentConverterTest.java
@@ -1,27 +1,31 @@
-package org.apiphany.header;
+package org.apiphany.client;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasSize;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
+import org.apiphany.ApiMessage;
+import org.apiphany.ApiMimeType;
+import org.apiphany.header.HeaderValues;
+import org.apiphany.header.MapHeaderValues;
 import org.apiphany.http.HttpHeaderValues;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.morphix.reflection.GenericClass;
 
 /**
- * Test class for {@link HeaderValues}.
+ * Test class for {@link ContentConverter}.
  *
  * @author Radu Sebastian LAZIN
  */
-class HeaderValuesTest {
+class ContentConverterTest {
 
 	private static final String HEADER_VALUE1 = "headerValue1";
 	private static final String HEADER_VALUE2 = "headerValue2";
@@ -38,29 +42,21 @@ class HeaderValuesTest {
 	@MethodSource("provideHeadersObjectsForChainWithoutHandlers")
 	void shouldReturnEmptyListFromHeadersWhenNoElementsAreAddedToTheChain(final Object headers) {
 		HeaderValues chain = new HeaderValues();
+		TestContentConverter converter = new TestContentConverter();
 
-		List<String> values = chain.get(HEADER_NAME1, headers);
+		List<String> values = converter.getHeaderValues(HEADER_NAME1, headers, chain);
 
 		assertThat(values, hasSize(0));
-	}
-
-	@ParameterizedTest
-	@MethodSource("provideHeadersObjectsForChainWithoutHandlers")
-	void shouldReturnFalseOnContainsWhenNoElementsAreAddedToTheChain(final Object headers) {
-		HeaderValues chain = new HeaderValues();
-
-		boolean contains = chain.contains(HEADER_NAME1, HEADER_VALUE1, headers);
-
-		assertFalse(contains);
 	}
 
 	@Test
 	void shouldReturnListFromMapHeadersChainContainsMapHeaderValues() {
 		HeaderValues chain = new HeaderValues().addFirst(new MapHeaderValues());
+		TestContentConverter converter = new TestContentConverter();
 
 		var headers = Map.of(HEADER_NAME1, List.of(HEADER_VALUE1));
 
-		List<String> values = chain.get(HEADER_NAME1, headers);
+		List<String> values = converter.getHeaderValues(HEADER_NAME1, headers, chain);
 
 		assertThat(values, hasSize(1));
 		assertThat(values.getFirst(), equalTo(HEADER_VALUE1));
@@ -69,13 +65,14 @@ class HeaderValuesTest {
 	@Test
 	void shouldReturnListFromMapHeadersWithMoreElementsChainContainsMapHeaderValues() {
 		HeaderValues chain = new HeaderValues().addFirst(new MapHeaderValues());
+		TestContentConverter converter = new TestContentConverter();
 
 		var headerValues = List.of(HEADER_VALUE1, HEADER_VALUE2);
 		var headers = Map.of(
 				HEADER_NAME1, headerValues,
 				"headerName2", List.of("v3", "v4"));
 
-		List<String> values = chain.get(HEADER_NAME1, headers);
+		List<String> values = converter.getHeaderValues(HEADER_NAME1, headers, chain);
 
 		assertThat(values, equalTo(headerValues));
 	}
@@ -84,9 +81,30 @@ class HeaderValuesTest {
 	@MethodSource("provideHeadersObjectsForChainWithoutHandlers")
 	void shouldReturnEmptyListFromHeadersWhenChainIsNotEmptyButNoHeaderValuesCanHandle(final Object headers) {
 		HeaderValues chain = new HeaderValues().addFirst(new HttpHeaderValues());
+		TestContentConverter converter = new TestContentConverter();
 
-		List<String> values = chain.get(HEADER_NAME1, headers);
+		List<String> values = converter.getHeaderValues(HEADER_NAME1, headers, chain);
 
 		assertThat(values, hasSize(0));
 	}
+
+	static class TestContentConverter implements ContentConverter<Object> {
+
+		@Override
+		public Object from(final Object obj, final ApiMimeType contentType, final Class<Object> dstClass) {
+			return null;
+		}
+
+		@Override
+		public Object from(final Object obj, final ApiMimeType contentType, final GenericClass<Object> genericDstClass) {
+			return null;
+		}
+
+		@Override
+		public <U, H> boolean isConvertible(final ApiMessage<U> message, final ApiMimeType contentType, final H headers, final HeaderValues headerValues) {
+			return false;
+		}
+
+	}
+
 }